+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.5] - 2024-01-28
+
+* Fixed list of printer status's to better align with what we expect from Bambu printers
+* Fixed light toggle
+* Fixed pause/resume/stop
+
 ## [1.0.4] - 2024-01-28
 
 * Added external spool support